+++ lib/transmuxer/aac_transmuxer.js
@@ -7,15 +7,10 @@
 goog.provide('shaka.transmuxer.AacTransmuxer');
 
 goog.require('shaka.media.Capabilities');
-goog.require('shaka.transmuxer.ADTS');
+goog.require('shaka.transmuxer.Mp4OutputTransmuxer');
 goog.require('shaka.transmuxer.TransmuxerEngine');
-goog.require('shaka.util.BufferUtils');
-goog.require('shaka.util.Error');
-goog.require('shaka.util.Id3Utils');
 goog.require('shaka.util.ManifestParserUtils');
 goog.require('shaka.util.MimeUtils');
-goog.require('shaka.util.Mp4Generator');
-goog.require('shaka.util.Uint8ArrayUtils');
 
 
 /**
@@ -30,11 +25,9 @@ shaka.transmuxer.AacTransmuxer = class {
     /** @private {string} */
     this.originalMimeType_ = mimeType;
 
-    /** @private {number} */
-    this.frameIndex_ = 0;
-
-    /** @private {!Map.<number, !Uint8Array>} */
-    this.initSegments = new Map();
+    /** @private {!shaka.transmuxer.Mp4OutputTransmuxer} */
+    this.mp4OutputTransmuxer = new shaka.transmuxer.Mp4OutputTransmuxer(
+        shaka.transmuxer.Mp4OutputTransmuxer.INPUT_FORMAT.AAC);
   }
 
 
@@ -43,7 +36,7 @@ shaka.transmuxer.AacTransmuxer = class {
    * @export
    */
   destroy() {
-    this.initSegments.clear();
+    this.mp4OutputTransmuxer.release();
   }
 
 
@@ -104,119 +97,9 @@ shaka.transmuxer.AacTransmuxer = class {
    * @override
    * @export
    */
-  transmux(data, stream, reference, duration) {
-    const ADTS = shaka.transmuxer.ADTS;
-    const Uint8ArrayUtils = shaka.util.Uint8ArrayUtils;
-
-    const uint8ArrayData = shaka.util.BufferUtils.toUint8(data);
-
-    // Check for the ADTS sync word
-    // Look for ADTS header | 1111 1111 | 1111 X00X | where X can be
-    // either 0 or 1
-    // Layer bits (position 14 and 15) in header should be always 0 for ADTS
-    // More info https://wiki.multimedia.cx/index.php?title=ADTS
-    const id3Data = shaka.util.Id3Utils.getID3Data(uint8ArrayData);
-    let offset = id3Data.length;
-    for (; offset < uint8ArrayData.length; offset++) {
-      if (ADTS.probe(uint8ArrayData, offset)) {
-        break;
-      }
-    }
-
-    let timestamp = reference.endTime * 1000;
-
-    const frames = shaka.util.Id3Utils.getID3Frames(id3Data);
-    if (frames.length && reference) {
-      const metadataTimestamp = frames.find((frame) => {
-        return frame.description ===
-            'com.apple.streaming.transportStreamTimestamp';
-      });
-      if (metadataTimestamp) {
-        timestamp = /** @type {!number} */(metadataTimestamp.data);
-      }
-    }
-
-    const info = ADTS.parseInfo(uint8ArrayData, offset);
-    if (!info) {
-      return Promise.reject(new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED));
-    }
-    stream.audioSamplingRate = info.sampleRate;
-    stream.channelsCount = info.channelCount;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    while (offset < uint8ArrayData.length) {
-      const header = ADTS.parseHeader(uint8ArrayData, offset);
-      if (!header) {
-        return Promise.reject(new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED));
-      }
-      const length = header.headerLength + header.frameLength;
-      if (offset + length <= uint8ArrayData.length) {
-        const data = uint8ArrayData.subarray(
-            offset + header.headerLength, offset + length);
-        samples.push({
-          data: data,
-          size: header.frameLength,
-          duration: ADTS.AAC_SAMPLES_PER_FRAME,
-          cts: 0,
-          flags: {
-            isLeading: 0,
-            isDependedOn: 0,
-            hasRedundancy: 0,
-            degradPrio: 0,
-            dependsOn: 2,
-            isNonSync: 0,
-          },
-        });
-      }
-      offset += length;
-    }
-
-    /** @type {number} */
-    const sampleRate = info.sampleRate;
-    /** @type {number} */
-    const baseMediaDecodeTime = Math.floor(timestamp * sampleRate / 1000);
-
-    /** @type {shaka.util.Mp4Generator.StreamInfo} */
-    const streamInfo = {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: info.codec,
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: new Uint8Array([]),
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-    const mp4Generator = new shaka.util.Mp4Generator([streamInfo]);
-    let initSegment;
-    if (!this.initSegments.has(stream.id)) {
-      initSegment = mp4Generator.initSegment();
-      this.initSegments.set(stream.id, initSegment);
-    } else {
-      initSegment = this.initSegments.get(stream.id);
-    }
-    const segmentData = mp4Generator.segmentData();
-
-    this.frameIndex_++;
-    const transmuxData = Uint8ArrayUtils.concat(initSegment, segmentData);
-    return Promise.resolve(transmuxData);
+  transmux(data, stream, reference, duration, contentType) {
+    return this.mp4OutputTransmuxer
+        .transmux(data, stream, reference, duration, contentType);
   }
 };
