+++ lib/transmuxer/ec3_transmuxer.js
@@ -7,14 +7,9 @@
 goog.provide('shaka.transmuxer.Ec3Transmuxer');
 
 goog.require('shaka.media.Capabilities');
-goog.require('shaka.transmuxer.Ec3');
+goog.require('shaka.transmuxer.Mp4OutputTransmuxer');
 goog.require('shaka.transmuxer.TransmuxerEngine');
-goog.require('shaka.util.BufferUtils');
-goog.require('shaka.util.Error');
-goog.require('shaka.util.Id3Utils');
 goog.require('shaka.util.ManifestParserUtils');
-goog.require('shaka.util.Mp4Generator');
-goog.require('shaka.util.Uint8ArrayUtils');
 
 
 /**
@@ -29,11 +24,9 @@ shaka.transmuxer.Ec3Transmuxer = class {
     /** @private {string} */
     this.originalMimeType_ = mimeType;
 
-    /** @private {number} */
-    this.frameIndex_ = 0;
-
-    /** @private {!Map.<number, !Uint8Array>} */
-    this.initSegments = new Map();
+    /** @private {!shaka.transmuxer.Mp4OutputTransmuxer} */
+    this.mp4OutputTransmuxer = new shaka.transmuxer.Mp4OutputTransmuxer(
+        shaka.transmuxer.Mp4OutputTransmuxer.INPUT_FORMAT.EC3);
   }
 
 
@@ -42,7 +35,7 @@ shaka.transmuxer.Ec3Transmuxer = class {
    * @export
    */
   destroy() {
-    this.initSegments.clear();
+    this.mp4OutputTransmuxer.release();
   }
 
 
@@ -102,110 +95,9 @@ shaka.transmuxer.Ec3Transmuxer = class {
    * @override
    * @export
    */
-  transmux(data, stream, reference, duration) {
-    const Ec3 = shaka.transmuxer.Ec3;
-    const Uint8ArrayUtils = shaka.util.Uint8ArrayUtils;
-
-    const uint8ArrayData = shaka.util.BufferUtils.toUint8(data);
-
-    const id3Data = shaka.util.Id3Utils.getID3Data(uint8ArrayData);
-    let offset = id3Data.length;
-    for (; offset < uint8ArrayData.length; offset++) {
-      if (Ec3.probe(uint8ArrayData, offset)) {
-        break;
-      }
-    }
-
-    let timestamp = reference.endTime * 1000;
-
-    const frames = shaka.util.Id3Utils.getID3Frames(id3Data);
-    if (frames.length && reference) {
-      const metadataTimestamp = frames.find((frame) => {
-        return frame.description ===
-            'com.apple.streaming.transportStreamTimestamp';
-      });
-      if (metadataTimestamp) {
-        timestamp = /** @type {!number} */(metadataTimestamp.data);
-      }
-    }
-
-    /** @type {number} */
-    let sampleRate = 0;
-
-    /** @type {!Uint8Array} */
-    let audioConfig = new Uint8Array([]);
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    while (offset < uint8ArrayData.length) {
-      const frame = Ec3.parseFrame(uint8ArrayData, offset);
-      if (!frame) {
-        return Promise.reject(new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED));
-      }
-      stream.audioSamplingRate = frame.sampleRate;
-      stream.channelsCount = frame.channelCount;
-      sampleRate = frame.sampleRate;
-      audioConfig = frame.audioConfig;
-
-      const frameData = uint8ArrayData.subarray(
-          offset, offset + frame.frameLength);
-
-      samples.push({
-        data: frameData,
-        size: frame.frameLength,
-        duration: Ec3.EC3_SAMPLES_PER_FRAME,
-        cts: 0,
-        flags: {
-          isLeading: 0,
-          isDependedOn: 0,
-          hasRedundancy: 0,
-          degradPrio: 0,
-          dependsOn: 2,
-          isNonSync: 0,
-        },
-      });
-      offset += frame.frameLength;
-    }
-    /** @type {number} */
-    const baseMediaDecodeTime = Math.floor(timestamp * sampleRate / 1000);
-
-    /** @type {shaka.util.Mp4Generator.StreamInfo} */
-    const streamInfo = {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: 'ec-3',
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: audioConfig,
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-    const mp4Generator = new shaka.util.Mp4Generator([streamInfo]);
-    let initSegment;
-    if (!this.initSegments.has(stream.id)) {
-      initSegment = mp4Generator.initSegment();
-      this.initSegments.set(stream.id, initSegment);
-    } else {
-      initSegment = this.initSegments.get(stream.id);
-    }
-    const segmentData = mp4Generator.segmentData();
-
-    this.frameIndex_++;
-    const transmuxData = Uint8ArrayUtils.concat(initSegment, segmentData);
-    return Promise.resolve(transmuxData);
+  transmux(data, stream, reference, duration, contentType) {
+    return this.mp4OutputTransmuxer
+        .transmux(data, stream, reference, duration, contentType);
   }
 };
