+++ lib/transmuxer/ts_transmuxer.js
@@ -7,23 +7,10 @@
 goog.provide('shaka.transmuxer.TsTransmuxer');
 
 goog.require('shaka.media.Capabilities');
-goog.require('shaka.transmuxer.Ac3');
-goog.require('shaka.transmuxer.ADTS');
-goog.require('shaka.transmuxer.Ec3');
-goog.require('shaka.transmuxer.H264');
-goog.require('shaka.transmuxer.H265');
-goog.require('shaka.transmuxer.MpegAudio');
+goog.require('shaka.transmuxer.Mp4OutputTransmuxer');
 goog.require('shaka.transmuxer.TransmuxerEngine');
-goog.require('shaka.util.BufferUtils');
-goog.require('shaka.util.Error');
-goog.require('shaka.util.Id3Utils');
 goog.require('shaka.util.ManifestParserUtils');
 goog.require('shaka.util.MimeUtils');
-goog.require('shaka.util.Mp4Generator');
-goog.require('shaka.util.TsParser');
-goog.require('shaka.util.Uint8ArrayUtils');
-
-goog.requireType('shaka.media.SegmentReference');
 
 
 /**
@@ -38,14 +25,9 @@ shaka.transmuxer.TsTransmuxer = class {
     /** @private {string} */
     this.originalMimeType_ = mimeType;
 
-    /** @private {number} */
-    this.frameIndex_ = 0;
-
-    /** @private {!Map.<number, !Uint8Array>} */
-    this.initSegments = new Map();
-
-    /** @private {?shaka.util.TsParser} */
-    this.tsParser_ = null;
+    /** @private {!shaka.transmuxer.Mp4OutputTransmuxer} */
+    this.mp4OutputTransmuxer = new shaka.transmuxer.Mp4OutputTransmuxer(
+        shaka.transmuxer.Mp4OutputTransmuxer.INPUT_FORMAT.TS);
   }
 
 
@@ -54,7 +36,7 @@ shaka.transmuxer.TsTransmuxer = class {
    * @export
    */
   destroy() {
-    this.initSegments.clear();
+    this.mp4OutputTransmuxer.release();
   }
 
 
@@ -187,841 +169,8 @@ shaka.transmuxer.TsTransmuxer = class {
    * @export
    */
   transmux(data, stream, reference, duration, contentType) {
-    const ContentType = shaka.util.ManifestParserUtils.ContentType;
-
-    const uint8ArrayData = shaka.util.BufferUtils.toUint8(data);
-
-    let transmuxData;
-    if (shaka.util.TsParser.probe(uint8ArrayData)) {
-      transmuxData = this.tsTransmuxing_(
-          uint8ArrayData, stream, reference, duration, contentType);
-    } else if (contentType == ContentType.AUDIO) {
-      transmuxData = this.aacTransmuxing_(
-          uint8ArrayData, stream, reference, duration);
-    }
-
-    if (!transmuxData) {
-      return Promise.reject(new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED));
-    }
-
-    return Promise.resolve(transmuxData);
-  }
-
-
-  /**
-   * @param {!Uint8Array} uint8ArrayData
-   * @param {shaka.extern.Stream} stream
-   * @param {?shaka.media.SegmentReference} reference
-   * @param {number} duration
-   * @param {string} contentType
-   * @return {!Uint8Array}
-   * @private
-   */
-  tsTransmuxing_(uint8ArrayData, stream, reference, duration, contentType) {
-    const ContentType = shaka.util.ManifestParserUtils.ContentType;
-    if (!this.tsParser_) {
-      this.tsParser_ = new shaka.util.TsParser();
-    } else {
-      this.tsParser_.clearData();
-    }
-
-    const tsParser = this.tsParser_.parse(uint8ArrayData);
-    const streamInfos = [];
-    const codecs = tsParser.getCodecs();
-    try {
-      let streamInfo = null;
-      if (contentType == ContentType.VIDEO) {
-        switch (codecs.video) {
-          case 'avc':
-            streamInfo =
-              this.getAvcStreamInfo_(tsParser, stream, duration, reference);
-            break;
-          case 'hvc':
-            streamInfo =
-              this.getHvcStreamInfo_(tsParser, stream, duration, reference);
-            break;
-        }
-        if (streamInfo) {
-          streamInfos.push(streamInfo);
-          streamInfo = null;
-        }
-      }
-      if (contentType == ContentType.AUDIO) {
-        switch (codecs.audio) {
-          case 'aac':
-            streamInfo =
-              this.getAacStreamInfo_(tsParser, stream, duration, reference);
-            break;
-          case 'ac3':
-            streamInfo =
-              this.getAc3StreamInfo_(tsParser, stream, duration);
-            break;
-          case 'ec3':
-            streamInfo =
-              this.getEc3StreamInfo_(tsParser, stream, duration);
-            break;
-          case 'mp3':
-            streamInfo =
-              this.getMp3StreamInfo_(tsParser, stream, duration);
-            break;
-        }
-        if (streamInfo) {
-          streamInfos.push(streamInfo);
-          streamInfo = null;
-        }
-      }
-    } catch (e) {
-      if (e && e.code == shaka.util.Error.Code.TRANSMUXING_NO_VIDEO_DATA) {
-        return new Uint8Array([]);
-      }
-      throw e;
-    }
-
-    if (!streamInfos.length) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-
-    const mp4Generator = new shaka.util.Mp4Generator(streamInfos);
-    let initSegment;
-    if (!this.initSegments.has(stream.id)) {
-      initSegment = mp4Generator.initSegment();
-      this.initSegments.set(stream.id, initSegment);
-    } else {
-      initSegment = this.initSegments.get(stream.id);
-    }
-    const segmentData = mp4Generator.segmentData();
-
-    this.frameIndex_++;
-    return shaka.util.Uint8ArrayUtils.concat(initSegment, segmentData);
-  }
-
-
-  /**
-   * @param {shaka.util.TsParser} tsParser
-   * @param {shaka.extern.Stream} stream
-   * @param {number} duration
-   * @param {?shaka.media.SegmentReference} reference
-   * @return {shaka.util.Mp4Generator.StreamInfo}
-   * @private
-   */
-  getAacStreamInfo_(tsParser, stream, duration, reference) {
-    const ADTS = shaka.transmuxer.ADTS;
-    const timescale = shaka.util.TsParser.Timescale;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    let info;
-
-    let firstPts = null;
-
-    for (const audioData of tsParser.getAudioData()) {
-      const data = audioData.data;
-      if (!data) {
-        continue;
-      }
-      let offset = 0;
-      info = ADTS.parseInfo(data, offset);
-      if (!info) {
-        throw new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED);
-      }
-      stream.audioSamplingRate = info.sampleRate;
-      stream.channelsCount = info.channelCount;
-      if (firstPts == null && audioData.pts !== null) {
-        firstPts = audioData.pts;
-      }
-      while (offset < data.length) {
-        const header = ADTS.parseHeader(data, offset);
-        if (!header) {
-          // We will increment one byte each time until we find the header.
-          offset++;
-          continue;
-        }
-        const length = header.headerLength + header.frameLength;
-        if (offset + length <= data.length) {
-          const frameData = data.subarray(
-              offset + header.headerLength, offset + length);
-
-          samples.push({
-            data: frameData,
-            size: header.frameLength,
-            duration: ADTS.AAC_SAMPLES_PER_FRAME,
-            cts: 0,
-            flags: {
-              isLeading: 0,
-              isDependedOn: 0,
-              hasRedundancy: 0,
-              degradPrio: 0,
-              dependsOn: 2,
-              isNonSync: 0,
-            },
-          });
-        }
-        offset += length;
-      }
-    }
-
-    if (!info || firstPts == null) {
-      if (!tsParser.getVideoData().length) {
-        throw new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED);
-      }
-
-      firstPts = reference.startTime * timescale;
-
-      const allCodecs = shaka.util.MimeUtils.splitCodecs(stream.codecs);
-      const audioCodec = shaka.util.ManifestParserUtils.guessCodecsSafe(
-          shaka.util.ManifestParserUtils.ContentType.AUDIO, allCodecs);
-      if (!audioCodec || !stream.channelsCount || !stream.audioSamplingRate) {
-        throw new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED);
-      }
-      info = {
-        sampleRate: stream.audioSamplingRate,
-        channelCount: stream.channelsCount,
-        codec: audioCodec,
-      };
-      const silenceFrame =
-          ADTS.getSilentFrame(audioCodec, stream.channelsCount);
-      if (!silenceFrame) {
-        throw new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED);
-      }
-      const segmentDuration =
-          (reference.endTime - reference.startTime) * timescale;
-      const finalPTs = firstPts + segmentDuration;
-      let currentPts = firstPts;
-      while (currentPts < finalPTs) {
-        samples.push({
-          data: silenceFrame,
-          size: silenceFrame.byteLength,
-          duration: ADTS.AAC_SAMPLES_PER_FRAME,
-          cts: 0,
-          flags: {
-            isLeading: 0,
-            isDependedOn: 0,
-            hasRedundancy: 0,
-            degradPrio: 0,
-            dependsOn: 2,
-            isNonSync: 0,
-          },
-        });
-        currentPts += ADTS.AAC_SAMPLES_PER_FRAME / info.sampleRate * timescale;
-      }
-    }
-
-    /** @type {number} */
-    const sampleRate = info.sampleRate;
-    /** @type {number} */
-    const baseMediaDecodeTime = firstPts / timescale * sampleRate;
-
-    return {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: info.codec,
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: new Uint8Array([]),
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-  }
-
-
-  /**
-   * @param {shaka.util.TsParser} tsParser
-   * @param {shaka.extern.Stream} stream
-   * @param {number} duration
-   * @return {shaka.util.Mp4Generator.StreamInfo}
-   * @private
-   */
-  getAc3StreamInfo_(tsParser, stream, duration) {
-    const Ac3 = shaka.transmuxer.Ac3;
-    const timescale = shaka.util.TsParser.Timescale;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    /** @type {number} */
-    let sampleRate = 0;
-
-    /** @type {!Uint8Array} */
-    let audioConfig = new Uint8Array([]);
-
-    let firstPts = null;
-
-    for (const audioData of tsParser.getAudioData()) {
-      const data = audioData.data;
-      if (firstPts == null && audioData.pts !== null) {
-        firstPts = audioData.pts;
-      }
-      let offset = 0;
-      while (offset < data.length) {
-        const frame = Ac3.parseFrame(data, offset);
-        if (!frame) {
-          offset++;
-          continue;
-        }
-        stream.audioSamplingRate = frame.sampleRate;
-        stream.channelsCount = frame.channelCount;
-        sampleRate = frame.sampleRate;
-        audioConfig = frame.audioConfig;
-
-        const frameData = data.subarray(
-            offset, offset + frame.frameLength);
-
-        samples.push({
-          data: frameData,
-          size: frame.frameLength,
-          duration: Ac3.AC3_SAMPLES_PER_FRAME,
-          cts: 0,
-          flags: {
-            isLeading: 0,
-            isDependedOn: 0,
-            hasRedundancy: 0,
-            degradPrio: 0,
-            dependsOn: 2,
-            isNonSync: 0,
-          },
-        });
-        offset += frame.frameLength;
-      }
-    }
-
-    if (sampleRate == 0 || audioConfig.byteLength == 0 || firstPts == null) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-
-
-    /** @type {number} */
-    const baseMediaDecodeTime = firstPts / timescale * sampleRate;
-
-    return {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: 'ac-3',
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: audioConfig,
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-  }
-
-
-  /**
-   * @param {shaka.util.TsParser} tsParser
-   * @param {shaka.extern.Stream} stream
-   * @param {number} duration
-   * @return {shaka.util.Mp4Generator.StreamInfo}
-   * @private
-   */
-  getEc3StreamInfo_(tsParser, stream, duration) {
-    const Ec3 = shaka.transmuxer.Ec3;
-    const timescale = shaka.util.TsParser.Timescale;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    /** @type {number} */
-    let sampleRate = 0;
-
-    /** @type {!Uint8Array} */
-    let audioConfig = new Uint8Array([]);
-
-    let firstPts = null;
-
-    for (const audioData of tsParser.getAudioData()) {
-      const data = audioData.data;
-      if (firstPts == null && audioData.pts !== null) {
-        firstPts = audioData.pts;
-      }
-      let offset = 0;
-      while (offset < data.length) {
-        const frame = Ec3.parseFrame(data, offset);
-        if (!frame) {
-          offset++;
-          continue;
-        }
-        stream.audioSamplingRate = frame.sampleRate;
-        stream.channelsCount = frame.channelCount;
-        sampleRate = frame.sampleRate;
-        audioConfig = frame.audioConfig;
-
-        const frameData = data.subarray(
-            offset, offset + frame.frameLength);
-
-        samples.push({
-          data: frameData,
-          size: frame.frameLength,
-          duration: Ec3.EC3_SAMPLES_PER_FRAME,
-          cts: 0,
-          flags: {
-            isLeading: 0,
-            isDependedOn: 0,
-            hasRedundancy: 0,
-            degradPrio: 0,
-            dependsOn: 2,
-            isNonSync: 0,
-          },
-        });
-        offset += frame.frameLength;
-      }
-    }
-
-    if (sampleRate == 0 || audioConfig.byteLength == 0 || firstPts == null) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-
-
-    /** @type {number} */
-    const baseMediaDecodeTime = firstPts / timescale * sampleRate;
-
-    return {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: 'ec-3',
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: audioConfig,
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-  }
-
-
-  /**
-   * @param {shaka.util.TsParser} tsParser
-   * @param {shaka.extern.Stream} stream
-   * @param {number} duration
-   * @return {shaka.util.Mp4Generator.StreamInfo}
-   * @private
-   */
-  getMp3StreamInfo_(tsParser, stream, duration) {
-    const MpegAudio = shaka.transmuxer.MpegAudio;
-    const timescale = shaka.util.TsParser.Timescale;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    let firstHeader;
-
-    let firstPts = null;
-
-    for (const audioData of tsParser.getAudioData()) {
-      const data = audioData.data;
-      if (!data) {
-        continue;
-      }
-      if (firstPts == null && audioData.pts !== null) {
-        firstPts = audioData.pts;
-      }
-      let offset = 0;
-      while (offset < data.length) {
-        const header = MpegAudio.parseHeader(data, offset);
-        if (!header) {
-          offset++;
-          continue;
-        }
-        if (!firstHeader) {
-          firstHeader = header;
-        }
-        if (offset + header.frameLength <= data.length) {
-          samples.push({
-            data: data.subarray(offset, offset + header.frameLength),
-            size: header.frameLength,
-            duration: MpegAudio.MPEG_AUDIO_SAMPLE_PER_FRAME,
-            cts: 0,
-            flags: {
-              isLeading: 0,
-              isDependedOn: 0,
-              hasRedundancy: 0,
-              degradPrio: 0,
-              dependsOn: 2,
-              isNonSync: 0,
-            },
-          });
-        }
-        offset += header.frameLength;
-      }
-    }
-    if (!firstHeader || firstPts == null) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-    /** @type {number} */
-    const sampleRate = firstHeader.sampleRate;
-    /** @type {number} */
-    const baseMediaDecodeTime = firstPts / timescale * sampleRate;
-
-    return {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: 'mp3',
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: new Uint8Array([]),
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-  }
-
-
-  /**
-   * @param {shaka.util.TsParser} tsParser
-   * @param {shaka.extern.Stream} stream
-   * @param {number} duration
-   * @param {?shaka.media.SegmentReference} reference
-   * @return {shaka.util.Mp4Generator.StreamInfo}
-   * @private
-   */
-  getAvcStreamInfo_(tsParser, stream, duration, reference) {
-    const H264 = shaka.transmuxer.H264;
-    const timescale = shaka.util.TsParser.Timescale;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    /** @type {?number} */
-    let baseMediaDecodeTime = null;
-
-    const nalus = [];
-    const videoData = tsParser.getVideoData();
-    if (!videoData.length) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_NO_VIDEO_DATA);
-    }
-    for (let i = 0; i < videoData.length; i++) {
-      const pes = videoData[i];
-      const dataNalus = pes.nalus;
-      nalus.push(...dataNalus);
-      const frame = H264.parseFrame(dataNalus);
-      if (!frame) {
-        continue;
-      }
-      if (baseMediaDecodeTime == null) {
-        baseMediaDecodeTime = pes.dts;
-      }
-      let duration;
-      if (i + 1 < videoData.length) {
-        duration = (videoData[i + 1].dts || 0) - (pes.dts || 0);
-      } else if (videoData.length > 1) {
-        duration = (pes.dts || 0) - (videoData[i - 1].dts || 0);
-      } else {
-        duration = (reference.endTime - reference.startTime) * timescale;
-      }
-      samples.push({
-        data: frame.data,
-        size: frame.data.byteLength,
-        duration: duration,
-        cts: Math.round((pes.pts || 0) - (pes.dts || 0)),
-        flags: {
-          isLeading: 0,
-          isDependedOn: 0,
-          hasRedundancy: 0,
-          degradPrio: 0,
-          dependsOn: frame.isKeyframe ? 2 : 1,
-          isNonSync: frame.isKeyframe ? 0 : 1,
-        },
-      });
-    }
-
-    const info = H264.parseInfo(nalus);
-
-    if (!info || baseMediaDecodeTime == null) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-    stream.height = info.height;
-    stream.width = info.width;
-
-    return {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.VIDEO,
-      codecs: 'avc1',
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: timescale,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: new Uint8Array([]),
-      videoConfig: info.videoConfig,
-      hSpacing: info.hSpacing,
-      vSpacing: info.vSpacing,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-  }
-
-
-  /**
-   * @param {shaka.util.TsParser} tsParser
-   * @param {shaka.extern.Stream} stream
-   * @param {number} duration
-   * @param {?shaka.media.SegmentReference} reference
-   * @return {shaka.util.Mp4Generator.StreamInfo}
-   * @private
-   */
-  getHvcStreamInfo_(tsParser, stream, duration, reference) {
-    const H265 = shaka.transmuxer.H265;
-    const timescale = shaka.util.TsParser.Timescale;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    /** @type {?number} */
-    let baseMediaDecodeTime = null;
-
-    const nalus = [];
-    const videoData = tsParser.getVideoData();
-    if (!videoData.length) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_NO_VIDEO_DATA);
-    }
-    for (let i = 0; i < videoData.length; i++) {
-      const pes = videoData[i];
-      const dataNalus = pes.nalus;
-      nalus.push(...dataNalus);
-      const frame = H265.parseFrame(dataNalus);
-      if (!frame) {
-        continue;
-      }
-      if (baseMediaDecodeTime == null && pes.dts != null) {
-        baseMediaDecodeTime = pes.dts;
-      }
-      let duration;
-      if (i + 1 < videoData.length) {
-        duration = (videoData[i + 1].dts || 0) - (pes.dts || 0);
-      } else if (videoData.length > 1) {
-        duration = (pes.dts || 0) - (videoData[i - 1].dts || 0);
-      } else {
-        duration = (reference.endTime - reference.startTime) * timescale;
-      }
-      samples.push({
-        data: frame.data,
-        size: frame.data.byteLength,
-        duration: duration,
-        cts: Math.round((pes.pts || 0) - (pes.dts || 0)),
-        flags: {
-          isLeading: 0,
-          isDependedOn: 0,
-          hasRedundancy: 0,
-          degradPrio: 0,
-          dependsOn: frame.isKeyframe ? 2 : 1,
-          isNonSync: frame.isKeyframe ? 0 : 1,
-        },
-      });
-    }
-
-    const info = H265.parseInfo(nalus);
-
-    if (!info || baseMediaDecodeTime == null) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-    stream.height = info.height;
-    stream.width = info.width;
-
-    return {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.VIDEO,
-      codecs: 'hvc1',
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: timescale,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: new Uint8Array([]),
-      videoConfig: info.videoConfig,
-      hSpacing: info.hSpacing,
-      vSpacing: info.vSpacing,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-  }
-
-  /** @private */
-  aacTransmuxing_(uint8ArrayData, stream, reference, duration) {
-    const ADTS = shaka.transmuxer.ADTS;
-    // Check for the ADTS sync word
-    // Look for ADTS header | 1111 1111 | 1111 X00X | where X can be
-    // either 0 or 1
-    // Layer bits (position 14 and 15) in header should be always 0 for ADTS
-    // More info https://wiki.multimedia.cx/index.php?title=ADTS
-    const id3Data = shaka.util.Id3Utils.getID3Data(uint8ArrayData);
-    let offset = id3Data.length;
-    for (; offset < uint8ArrayData.length; offset++) {
-      if (ADTS.probe(uint8ArrayData, offset)) {
-        break;
-      }
-    }
-
-    let timestamp = reference.endTime * 1000;
-
-    const frames = shaka.util.Id3Utils.getID3Frames(id3Data);
-    if (frames.length && reference) {
-      const metadataTimestamp = frames.find((frame) => {
-        return frame.description ===
-            'com.apple.streaming.transportStreamTimestamp';
-      });
-      if (metadataTimestamp) {
-        timestamp = /** @type {!number} */(metadataTimestamp.data);
-      }
-    }
-
-    const info = ADTS.parseInfo(uint8ArrayData, offset);
-    if (!info) {
-      throw new shaka.util.Error(
-          shaka.util.Error.Severity.CRITICAL,
-          shaka.util.Error.Category.MEDIA,
-          shaka.util.Error.Code.TRANSMUXING_FAILED);
-    }
-    stream.audioSamplingRate = info.sampleRate;
-    stream.channelsCount = info.channelCount;
-
-    /** @type {!Array.<shaka.util.Mp4Generator.Mp4Sample>} */
-    const samples = [];
-
-    while (offset < uint8ArrayData.length) {
-      const header = ADTS.parseHeader(uint8ArrayData, offset);
-      if (!header) {
-        throw new shaka.util.Error(
-            shaka.util.Error.Severity.CRITICAL,
-            shaka.util.Error.Category.MEDIA,
-            shaka.util.Error.Code.TRANSMUXING_FAILED);
-      }
-      const length = header.headerLength + header.frameLength;
-      if (offset + length <= uint8ArrayData.length) {
-        const data = uint8ArrayData.subarray(
-            offset + header.headerLength, offset + length);
-        samples.push({
-          data: data,
-          size: header.frameLength,
-          duration: ADTS.AAC_SAMPLES_PER_FRAME,
-          cts: 0,
-          flags: {
-            isLeading: 0,
-            isDependedOn: 0,
-            hasRedundancy: 0,
-            degradPrio: 0,
-            dependsOn: 2,
-            isNonSync: 0,
-          },
-        });
-      }
-      offset += length;
-    }
-
-    /** @type {number} */
-    const sampleRate = info.sampleRate;
-    /** @type {number} */
-    const baseMediaDecodeTime = Math.floor(timestamp * sampleRate / 1000);
-
-    /** @type {shaka.util.Mp4Generator.StreamInfo} */
-    const streamInfo = {
-      id: stream.id,
-      type: shaka.util.ManifestParserUtils.ContentType.AUDIO,
-      codecs: info.codec,
-      encrypted: stream.encrypted && stream.drmInfos.length > 0,
-      timescale: sampleRate,
-      duration: duration,
-      videoNalus: [],
-      audioConfig: new Uint8Array([]),
-      videoConfig: new Uint8Array([]),
-      hSpacing: 0,
-      vSpacing: 0,
-      data: {
-        sequenceNumber: this.frameIndex_,
-        baseMediaDecodeTime: baseMediaDecodeTime,
-        samples: samples,
-      },
-      stream: stream,
-    };
-    const mp4Generator = new shaka.util.Mp4Generator([streamInfo]);
-    let initSegment;
-    if (!this.initSegments.has(stream.id)) {
-      initSegment = mp4Generator.initSegment();
-      this.initSegments.set(stream.id, initSegment);
-    } else {
-      initSegment = this.initSegments.get(stream.id);
-    }
-    const segmentData = mp4Generator.segmentData();
-
-    this.frameIndex_++;
-    return shaka.util.Uint8ArrayUtils.concat(initSegment, segmentData);
+    return this.mp4OutputTransmuxer
+        .transmux(data, stream, reference, duration, contentType);
   }
 };
