+++ shaka-player.uncompiled.js
@@ -73,6 +73,7 @@ goog.require('shaka.transmuxer.Ec3Transmuxer');
 goog.require('shaka.transmuxer.H264');
 goog.require('shaka.transmuxer.H265');
 goog.require('shaka.transmuxer.Mp3Transmuxer');
+goog.require('shaka.transmuxer.Mp4OutputTransmuxer');
 goog.require('shaka.transmuxer.MpegAudio');
 goog.require('shaka.transmuxer.MpegTsTransmuxer');
 goog.require('shaka.transmuxer.TransmuxerEngine');
